+++ sdk/messaging/azeventgrid/example_test.go
@@ -57,7 +57,9 @@ func ExampleClient_PublishCloudEvents() {
 	var eventsToSend []messaging.CloudEvent
 
 	for _, payload := range payloads {
-		event, err := messaging.NewCloudEvent("source", "eventType", payload, nil)
+		event, err := messaging.NewCloudEvent("source", "eventType", payload, &messaging.CloudEventOptions{
+			DataContentType: to.Ptr("application/octet-stream"),
+		})
 
 		if err != nil {
 			//  TODO: Update the following line with your application specific error handling logic