+++ sdk/messaging/azeventgrid/main_test.go
@@ -7,11 +7,20 @@
 package azeventgrid_test
 
 import (
+	"context"
+	"fmt"
 	"log"
+	"net/url"
 	"os"
+	"strings"
 	"testing"
+	"time"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/recording"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/eventgrid/armeventgrid/v2"
 	"github.com/joho/godotenv"
 )
 
@@ -40,5 +49,88 @@ func run(m *testing.M) int {
 		log.Printf("Failed to load .env file, no integration tests will run: %s", err)
 	}
 
+	azSubID := os.Getenv("AZEVENTGRID_SUBSCRIPTION_ID")
+	resGroup := os.Getenv("AZEVENTGRID_RESOURCE_GROUP")
+
+	nsURL, err := url.Parse(os.Getenv("EVENTGRID_ENDPOINT"))
+
+	if err != nil {
+		panic(err)
+	}
+
+	nsHost := strings.Split(nsURL.Host, ".")[0]
+
+	//cred, err := azidentity.NewAzureCLICredential(nil)
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+
+	if err != nil {
+		panic(err)
+	}
+
+	topicClient, err := armeventgrid.NewNamespaceTopicsClient(azSubID, cred, nil)
+
+	if err != nil {
+		panic(err)
+	}
+
+	subClient, err := armeventgrid.NewNamespaceTopicEventSubscriptionsClient(azSubID, cred, nil)
+
+	if err != nil {
+		panic(err)
+	}
+
+	topicName := fmt.Sprintf("topic-%d", time.Now().UnixNano())
+
+	os.Setenv("EVENTGRID_TOPIC", topicName)
+
+	subName := "testsubscription1"
+
+	err = PollUntilDone(context.Background(), func() (*runtime.Poller[armeventgrid.NamespaceTopicsClientCreateOrUpdateResponse], error) {
+		return topicClient.BeginCreateOrUpdate(context.Background(), resGroup, nsHost, topicName, armeventgrid.NamespaceTopic{}, nil)
+	})
+
+	if err != nil {
+		panic(err)
+	}
+
+	defer func() {
+		if _, err = topicClient.BeginDelete(context.Background(), resGroup, nsHost, topicName, nil); err != nil {
+			fmt.Printf("Failed to start the delete for our test topic %s: %s", topicName, err)
+		}
+	}()
+
+	err = PollUntilDone(context.Background(), func() (*runtime.Poller[armeventgrid.NamespaceTopicEventSubscriptionsClientCreateOrUpdateResponse], error) {
+		return subClient.BeginCreateOrUpdate(context.Background(),
+			resGroup,
+			nsHost,
+			topicName,
+			subName,
+			armeventgrid.Subscription{
+				Properties: &armeventgrid.SubscriptionProperties{
+					DeliveryConfiguration: &armeventgrid.DeliveryConfiguration{
+						DeliveryMode: to.Ptr(armeventgrid.DeliveryModeQueue),
+					},
+				},
+			},
+			nil)
+	})
+
+	if err != nil {
+		panic(err)
+	}
+
+	fmt.Printf("Created topic %s\n", topicName)
+
 	return m.Run()
+}
+
+func PollUntilDone[T any](ctx context.Context, fn func() (*runtime.Poller[T], error)) error {
+	poller, err := fn()
+
+	if err != nil {
+		return err
+	}
+
+	_, err = poller.PollUntilDone(ctx, nil)
+	return err
 }