+++ sdk/messaging/azeventgrid/CHANGELOG.md
@@ -1,10 +1,19 @@
 # Release History
 
-## 0.3.1 (Unreleased)
+## 0.4.0 (2023-11-17)
 
 ### Features Added
 
+- New functionality for Event Grid namespaces: 
+  - Client.PublishCloudEvent can be used to publish a single `messaging.CloudEvent`.
+  - Client.RenewCloudEventLocks can extend the lock time for a set of events.
+  - Client.ReleaseCloudEvents (via ReleaseCloudEventsOptions.ReleaseDelayInSeconds) can release an event with a 
+    server-side delay, allowing the message to remain unavailable for a configured period of time.
+
 ### Breaking Changes
+
+- FailedLockToken, included in the response for settlement functions, has an `Error` field, which contains the data previously
+  in `ErrorDescription` and `ErrorCode`.
 
 ### Bugs Fixed